requirements.txt
streamlit==1.11.1
nltk==3.8.1
numpy==1.24.2
pandas==1.5.3
protobuf==3.19.5
urllib3==1.26.14
matplotlib==3.7.1
matplotlib-inline==0.1.6
google-api-python-client==2.95.0
wordcloud==1.9.2
beautifulsoup4